Churn Prediction in Telco/deployment/requirements.txt
streamlit
pandas
numpy
tensorflow-cpu
scikit-learn